refactor(editor): Extract node issue handling to separate composable and utilities
Move node issue state management logic from useNodeHelpers to a dedicated useWorkflowNodeIssuesState composable and extract utility functions into nodeIssueUtils. This improves code organization and separation of concerns.
