+++ packages/frontend/editor-ui/src/app/components/FocusPanel.vue
@@ -155,7 +155,7 @@ const isExecutable = computed(() => {
 
 	if (!isDisplayed.value) return false;
 
-	const foreignCredentials = nodeHelpers.getForeignCredentialsIfSharingEnabled(
+	const foreignCredentials = workflowState.getForeignCredentialsIfSharingEnabled(
 		node.value.credentials,
 	);
 	return nodeHelpers.isNodeExecutable(node.value, !props.isCanvasReadOnly, foreignCredentials);
