+++ packages/frontend/editor-ui/src/app/composables/useWorkflowUpdate.ts
@@ -11,7 +11,6 @@ import { useNodeTypesStore } from '@/app/stores/nodeTypes.store';
 import { useBuilderStore } from '@/features/ai/assistant/builder.store';
 import { injectWorkflowState } from '@/app/composables/useWorkflowState';
 import { useCanvasOperations } from '@/app/composables/useCanvasOperations';
-import { useNodeHelpers } from '@/app/composables/useNodeHelpers';
 import { canvasEventBus } from '@/features/workflows/canvas/canvas.eventBus';
 import { mapLegacyConnectionsToCanvasConnections } from '@/features/workflows/canvas/canvas.utils';
 import { getAuthTypeForNodeCredential, getMainAuthField } from '@/app/utils/nodeTypesUtils';
@@ -43,7 +42,6 @@ export function useWorkflowUpdate() {
 	const nodeTypesStore = useNodeTypesStore();
 	const builderStore = useBuilderStore();
 	const canvasOperations = useCanvasOperations();
-	const nodeHelpers = useNodeHelpers();
 
 	/**
 	 * Categorize nodes into those to update, add, or remove
@@ -128,7 +126,7 @@ export function useWorkflowUpdate() {
 		// Revalidate updated nodes to refresh error indicators on canvas
 		for (const { existing } of nodesToUpdate) {
 			const nodeName = renamedNodes.get(existing.id) ?? existing.name;
-			nodeHelpers.updateNodeParameterIssuesByName(nodeName);
+			workflowState.updateNodeParameterIssuesByName(nodeName);
 		}
 	}
