+++ packages/frontend/editor-ui/src/features/execution/logs/composables/useLogsExecutionData.ts
@@ -2,7 +2,6 @@ import { watch, computed, ref, type ComputedRef } from 'vue';
 import type { IExecutionResponse } from '@/features/execution/executions/executions.types';
 import { Workflow, type IRunExecutionData, type ITaskStartedData } from 'n8n-workflow';
 import { useWorkflowsStore } from '@/app/stores/workflows.store';
-import { useNodeHelpers } from '@/app/composables/useNodeHelpers';
 import {
 	copyExecutionData,
 	createLogTree,
@@ -27,7 +26,6 @@ interface UseLogsExecutionDataOptions {
 }
 
 export function useLogsExecutionData({ isEnabled, filter }: UseLogsExecutionDataOptions = {}) {
-	const nodeHelpers = useNodeHelpers();
 	const workflowsStore = useWorkflowsStore();
 	const workflowState = injectWorkflowState();
 	const toast = useToast();
@@ -92,7 +90,7 @@ export function useLogsExecutionData({ isEnabled, filter }: UseLogsExecutionData
 	function resetExecutionData() {
 		state.value = undefined;
 		workflowState.setWorkflowExecutionData(null);
-		nodeHelpers.updateNodesExecutionIssues();
+		workflowState.updateNodesExecutionIssues();
 		void workflowsStore.fetchLastSuccessfulExecution();
 	}
