+++ packages/frontend/editor-ui/src/features/ndv/panel/components/NDVSubConnections.vue
@@ -4,7 +4,8 @@ import { useNodeTypesStore } from '@/app/stores/nodeTypes.store';
 import { useWorkflowsStore } from '@/app/stores/workflows.store';
 import { computed, ref, watch } from 'vue';
 import { NodeHelpers } from 'n8n-workflow';
-import { useNodeHelpers } from '@/app/composables/useNodeHelpers';
+import { injectWorkflowState } from '@/app/composables/useWorkflowState';
+import { nodeIssuesToString } from '@/app/utils/nodeIssueUtils';
 import NodeIcon from '@/app/components/NodeIcon.vue';
 import TitledList from '@/app/components/TitledList.vue';
 import type {
@@ -26,7 +27,7 @@ interface Props {
 const props = defineProps<Props>();
 const workflowsStore = useWorkflowsStore();
 const nodeTypesStore = useNodeTypesStore();
-const nodeHelpers = useNodeHelpers();
+const workflowState = injectWorkflowState();
 const i18n = useI18n();
 const { debounce } = useDebounce();
 const emit = defineEmits<{
@@ -65,7 +66,7 @@ const ndvStore = useNDVStore();
 const workflowObject = computed(() => workflowsStore.workflowObject as Workflow);
 
 const nodeInputIssues = computed(() => {
-	const issues = nodeHelpers.getNodeIssues(nodeType.value, props.rootNode, workflowObject.value, [
+	const issues = workflowState.getNodeIssues(nodeType.value, props.rootNode, workflowObject.value, [
 		'typeUnknown',
 		'parameters',
 		'credentials',
@@ -163,8 +164,8 @@ function getINodesFromNames(names: string[]): NodeConfig[] {
 			if (node) {
 				const matchedNodeType = nodeTypesStore.getNodeType(node.type);
 				if (matchedNodeType) {
-					const issues = nodeHelpers.getNodeIssues(matchedNodeType, node, workflowObject.value);
-					const stringifiedIssues = issues ? nodeHelpers.nodeIssuesToString(issues, node) : '';
+					const issues = workflowState.getNodeIssues(matchedNodeType, node, workflowObject.value);
+					const stringifiedIssues = issues ? nodeIssuesToString(issues, node) : '';
 					return { node, nodeType: matchedNodeType, issues: stringifiedIssues };
 				}
 			}
